migrate from juntagricoapp.py to apps.py
