--- juntagrico_assignment_request/templates/assignment_request/menu/assignment_request_admin_menu.html
@@ -1,8 +1,0 @@
-{% load assignment_request_config %}
-{% load i18n %}
-<li class="nav-item">
-    {% vocabulary "assignment_pl" as v_assignment_pl %}
-    <a class="nav-link" href="{% url 'juntagrico-assignment-request:list' %}">
-        {% blocktrans %}Angemeldete {{ v_assignment_pl }}{% endblocktrans %}
-    </a>
-</li>
