--- juntagrico_assignment_request/templates/assignment_request/menu/assignment_request_user_menu.html
@@ -1,7 +1,0 @@
-{% load juntagrico.common %}
-{% load assignment_request_config %}
-{% load i18n %}
-<li class="nav-item">
-    {% vocabulary "assignment" as v_assignment %}
-    <a class="nav-link {{ request|activemenu:"request_assignment" }}" href="{% url 'juntagrico-assignment-request:request' %}">{% blocktrans %}{{ v_assignment }} melden{% endblocktrans %}</a>
-</li>
