+++ juntagrico_assignment_request/templates/menu/admin_menu.html
@@ -1,0 +1,12 @@
+{% extends 'menu/admin_menu.html' %}
+{% load i18n %}
+{% load juntagrico.config %}
+{% block extend_admin_menu %}
+    <li class="nav-item">
+        {% vocabulary "assignment_pl" as v_assignment_pl %}
+        <a class="nav-link" href="{% url 'juntagrico-assignment-request:list' %}">
+            {% blocktrans %}Angemeldete {{ v_assignment_pl }}{% endblocktrans %}
+        </a>
+    </li>
+    {{ block.super }}
+{% endblock %}
