+++ juntagrico_assignment_request/templates/menu/user_menu.html
@@ -1,0 +1,14 @@
+{% extends 'menu/user_menu.html' %}
+{% load i18n %}
+{% load juntagrico.common %}
+{% load juntagrico.config %}
+{% block extend_user_menu %}
+    <li class="nav-item">
+        {% vocabulary "assignment" as v_assignment %}
+        <a class="nav-link {{ request|activemenu:"request_assignment" }}" 
+           href="{% url 'juntagrico-assignment-request:request' %}">
+            {% blocktrans %}{{ v_assignment }} melden{% endblocktrans %}
+        </a>
+    </li>
+    {{ block.super }}
+{% endblock %}
